Extract catalog axis selection and data type utilities from CatalogOverlayComponent into standalone utility module with comprehensive test coverage
