+++ src/utilities/index.ts
@@ -1,5 +1,6 @@
 export * from "./array/array";
 export * from "./ast/ast";
+export * from "./catalog/catalog";
 export * from "./CatalogApiProcessed/CatalogApiProcessed";
 export * from "./color/color";
 export * from "./export/export";
